pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wmoov-weekend-scraper"
version = "0.1.0"
description = "A Playwright-based scraper for WMOOV weekend movies"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "playwright>=1.40.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "python-dateutil>=2.8.2",
    "beautifulsoup4>=4.12.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
wmoov-scraper = "wmoov_scraper.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["wmoov_scraper"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"